requirements.txt
# required for Interacting with Telegram API
python-telegram-bot

# required for calling the THIRD PARTY APIs
requests
